+++ CHANGES.md
@@ -6,6 +6,11 @@ This files lists the changes during the lifetime of this project.
 
 - 324: replace raw HTML with JRC components across templates and include roos.css
 - 324: add sortable Tot column to placement table
+
+## 2026-05-05
+
+- 319: fix user import non-utf8
+- 319: fix user import check existence case insensitive
 - 304: (migration) register assignment edit events and display on 'Updates' tab
 - 304: change dummy data to 50/50 split wies/otys sources
 - 304: bump jinja-roos-components to 0.5